Created a new branch of the CollabTest branch
In this test branch, the system can process a user's request to display
the correct number of games in the game select screen (the bonus level
games cannot be seen just yet). However, the game options are not
displaying with the heights that were intended.  I have added a main
controller to the system, and propose that a shell module be created to
handle basic app functions, such as page navigation (see nav.js). The
viewModel is more of a view controller in this branch.