Fixed a bug in the game
The validateWords() function was not using the correct index to remove
the word which the user had chosen. To make it easier to read the
functions in mainController I changed the names of some of the
functions. The system can save data to allGamesData once the user has
finished a game. There is a bug which occurs after the user has finished
the game