Working on continue
The problem is if the user returns to a previous game, then the system
only shows the words (in cells) which the user has not answered for that
game. Also, copies of the game's wordlist array are becoming empty for
some strange reason. Even the currentGameData.myGame.myWordList is
sometimes empty.