Fixed the continue button bug
When the user clicks the Continue button to return to a previous game,
the table on the game screen will now show the full number of words
which are in the game's word-list. The solution is to use slice(0) every
time the word list is copied from the game's word-list. It's still hard
to understand how it works, for instance the completeWordList is an
empty array when I look at it during the game.