Continue btn & changes to setting a new player
If the player is on the home screen and does not have any saved games
and clicks the continue button then the game which matches the highest
level-game the player has achieved will be opened. Question - should the
continue button be able to do this? Or should it just be disabled if the
user has no partially completed games? Also, the way players are added
to the system has been changed because previously the objects which were
assigned to rocketReadingModel.myPlayer were not Player class objects.