Fixed game-test timer
If a player leaves a game and then returns to it later, the timer for
the last test of the previous had been kept in the system and would be
used for the first test of the player's resumption of the game. Now, the
game-test timer is cleared when a player leaves the current game.