Decreased the padding-left of the medalIconDivs
of the game screen because the former setting was causing the divs to
break when the screen was full-screen (on the monitors at CPIT).
However, now the medals are perhaps too far to the left .... Also, added
'use strict' to some functions. Question - at some point are we going to
have to make the program JSLint compliant?