Implemented part of a learn word scenario -
If the user makes a wrong choice, and the user clicks the red-coloured
'Learn Word' div then the correct word will be spoken twice. I initially
tried to make 'learnWordCount' an input variable of updateCurrentWord,
however I found that I had to make it a global var to get it to work.