Learn Word mode functionality
If the user gets the word he or she is learning incorrect more than
once, then the number of words on display should be reduced by a third
of the original length of the word-list. However, the system still isn't
working as well as it should. The problem lies with the randomiseArray()
function: it creates a new array in which its items are randomly
arranged. However, it results in the array being passed to it
effectively being wiped. This causes problems because the array which is
being passed to this randomiseArray() could potentially be passed more
than once to validateWords().