Learn word dotted text
The learn word sequence will display the current word in a dotted custom
font at the appropriate time in the learn word sequence. I have put the
JarmanDotted font on Github. However, it is not possible to install it
on the computers at CPIT because of the permissions issue (I guess).