Incorrect words / sounds
The system was adding 'null' items to the wordsSoundsIncorrect array
property of currentGameData if a user got a word incorrect and then
later correctly chose the right word. I have created a separate function
which will only be called when a word/sound needs to be added to the
wordsSoundsIncorrect array.