The words in word-lists are no longer randomised
before being displayed on the game screen's table. They will now display
in the order in which they are found in the original word-list (which is
what the client wants). Also, I have removed a css bug in the
Register_layout css file - which was affecting all p tags of the app.