accessing all the pages
If you want to test all the pages,
set up a user using rocketReadingModel.registerPlayer. it allows you to
specify the levelGameReached, and thereby give access to every game.
I have been tampering around trying to deal with localstorage. Having
issues with saving allGamesData into localstorage (this is needed,
because when you login, you need to have allGamesData in memory to play
the game)