If a user leaves a game before the user has
finished the game and then later returns to that game, the current word
will be the same word which the user was being tested on when they
previously left the game. Furthermore, if the user was having to learn a
word when they left the game, they will have to learn that same word
when they return to the game.