When the player finishes a game the
it seems that currentGameData needs to be saved to local storage, even
if it is empty, in order to allow the player's data to be loaded when
the browser is refreshed and the program is reloaded - there needs to be
a currentGameData object in the player's file in LS, even if it is
empty.
Also, tidied up the currentGameData file a bit.