I have split setAccessTo() into two functions
The new function, setLevelGameReached(), only needs to be called when
the user has finished the game. Still trying to fix the bug which
prevents the user from unlocking the next level's game 1 when the
prereq.s for this have been satisfied. A problem: getting the system to
remember that the player has completed a bonus game. Should this sort of
data be saved to the allGamesData property of myPlayer (because it would
be permanently stored there and would be untouched)?