Trying to fix levelGameReached bug.
Problem: if the user completes a bonus game the system sets the
bonusGameCompleted property of this game in currentGameData to true.
However, if the user changes level then this information will be lost.
Probably will have to save the data in a allGamesData property of the
myPlayer, so t hat it will be persistent data. I have changed your code
a bit, Nathan. Feel free to change it back to how it was if you don't
like it.