Player can now leave a partly completed game
and when they login in later, they can return to that last game. Had to
change code which referenced the myLevel of the currentGameData object -
because when a player logs out the myLevel property cannot be saved
(causes circular JSON error). Thankfully, it is possible to get data
about the level by using the currentLevelGame of currentGameData.