Adding functions to mainController module
- relating to the starting and replaying of games in order to enable a
new game to be displayed on a screen without necessarily clobbering an
old game that the player may have played. On the game intro modal screen
if the player clicks the start button a new game really will be started.
However, if the player clicks the back button then the player could
potentially go back to playing their old game. There are still a couple
of issues to sort out with the code.