Continue functionality
If a learn word sequence has been activated and the user leaves the
current game and comes back later - should the user have to go through
the learn word sequence? (I guess so, just for the sake of the
consistency of the scoring of a game.)  Anyway, I have fixed up the
continue game functionality.