Continue game functionality
A player can leave their current game and navigate to the play another
game. Just as long as the player does not click the start button in the
game-intro modal window of the the new game, then the player can now
return to the home page and can click the continue button to carry on
with their previously unfinished game.