Improved navigation and game saving
- The event listeners which will be added to navigation buttons,
including buttons in the modal windows, is now determined by the screen
/ state the user is currently in.
- Saved games will still exist even if the user navigates to start a new
game, but does not actually click the start button to begin a new game.
- A saved game data property has been added to the Player class. If a
player leaves a game before it has been finished, then the data for that
game will be saved to this new property. If the user decides to continue
playing the saved game, then the data in the saved game property will be
loaded into the currentGameData object, and the user can keep playing
where they left off. Although this was possible prior to this addition,
it means that if a user leaves a game and navigates to start a new game
then the data from the old game will not be displayed behind the
gameIntro modal for the new game.
- When a user registers a new account with the app, spaces before and
after data the user enters will be trimmed off.